// src/generated/types/StudentGradesSummaryDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { EventMarkDto } from './EventMarkDto.ts'

export type StudentGradesSummaryDto = {
  /**
   * @type string
   */
  studentId: string
  /**
   * @type string
   */
  studentName: string
  /**
   * @type string
   */
  username: string
  /**
   * @type string
   */
  seatNo: string
  /**
   * @type number
   */
  totalMarks: number
  /**
   * @type number
   */
  earnedMarks: number
  /**
   * @type number
   */
  percentage: number
  /**
   * @type array
   */
  eventMarks: EventMarkDto[]
}